Refactor packages, introduce deprecated InfinityScreen
Hmm, thinking about whether I even need this
