+++ src/main/java/com/github/frontear/infinity/InfinityMod.java
@@ -3,10 +3,13 @@ package com.github.frontear.infinity;
 import com.github.frontear.efkolia.impl.mod.MinecraftMod;
 import com.github.frontear.infinity.commands.CommandContainer;
 import com.github.frontear.infinity.modules.ModuleContainer;
+import com.github.frontear.infinity.ux.TextRenderer;
 import com.google.gson.JsonObject;
 import lombok.*;
+import net.minecraft.client.MinecraftClient;
 
 public final class InfinityMod extends MinecraftMod {
+    @Getter private final TextRenderer renderer;
     @Getter private final ModuleContainer modules;
     @Getter private final CommandContainer commands;
 
@@ -16,5 +19,6 @@ public final class InfinityMod extends MinecraftMod {
 
         this.modules = new ModuleContainer(this);
         this.commands = new CommandContainer(this);
+        this.renderer = new TextRenderer(this, MinecraftClient.getInstance());
     }
 }
