+++ src/main/java/com/github/frontear/infinity/commands/CommandContainer.java
@@ -3,6 +3,7 @@ package com.github.frontear.infinity.commands;
 import com.github.frontear.efkolia.api.events.Listener;
 import com.github.frontear.efkolia.impl.container.Container;
 import com.github.frontear.infinity.InfinityMod;
+import com.github.frontear.infinity.commands.ux.CommandScreen;
 import com.github.frontear.infinity.event.input.KeyEvent;
 import com.github.frontear.infinity.modules.impl.Ghost;
 import lombok.NonNull;
